% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasty.R
\name{pasty}
\alias{pasty}
\title{Create a character vector of capture histories}
\usage{
pasty(x)
}
\arguments{
\item{x}{A dataframe with rows of individuals and columns of capture events}
}
\value{
A character vector
}
\description{
Copied from somewhere
}
