% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simul_js.R
\name{simul_js}
\alias{simul_js}
\title{A function to simulate open population capture-recapture data}
\usage{
simul_js(PHI, P, b, N)
}
\arguments{
\item{PHI}{A n by k - 1 array of apparent survivals where n is population size and k is the number of events}

\item{P}{A n by k array of probabilities of capture where n is population size and k is the number of events}

\item{b}{A k length vector of entry probabilities where k is the number of events}

\item{N}{population size}
}
\value{
A list with a character vector of capture histories, a vector of new entrants and a vector of populations size.
}
\description{
Copied from Kery and Schaub BPA pg. 329, modified so that CH is in MARK format
}
